requirements.txt
opencv-contrib-python
numpy
Pillow
future
pygame